+++ src/main/java/org/codinjutsu/tools/jenkins/view/action/UnsetJobAsFavoriteAction.java
@@ -17,11 +17,13 @@
 package org.codinjutsu.tools.jenkins.view.action;
 
 import com.intellij.icons.AllIcons;
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.DumbAware;
 import org.codinjutsu.tools.jenkins.model.Job;
 import org.codinjutsu.tools.jenkins.view.BrowserPanel;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
@@ -44,5 +46,10 @@ public class UnsetJobAsFavoriteAction extends AnAction implements DumbAware {
     public void update(AnActionEvent event) {
         Job selectedJob = browserPanel.getSelectedJob();
         event.getPresentation().setVisible(selectedJob != null && browserPanel.isAFavoriteJob(selectedJob));
+    }
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
     }
 }